+++ catalog/app/utils/BucketPreferences/BucketPreferences.ts
@@ -1,10 +1,11 @@
+import type { ErrorObject } from 'ajv'
 import * as R from 'ramda'
 import * as Sentry from '@sentry/react'
 
 import bucketPreferencesSchema from 'schemas/bucketConfig.yml.json'
 
-import * as bucketErrors from 'containers/Bucket/errors'
 import { makeSchemaValidator } from 'utils/JSONSchema'
+import { JsonInvalidAgainstSchema } from 'utils/error'
 import * as tagged from 'utils/taggedV2'
 import type { JsonRecord } from 'utils/types'
 import * as YAML from 'utils/yaml'
@@ -21,7 +22,7 @@ export type ActionPreferences = Record<
   boolean
 >
 
-interface MetaBlockPreferencesInput {
+export interface MetaBlockPreferencesInput {
   user_meta?: {
     expanded: boolean | number
   }
@@ -61,8 +62,8 @@ interface BlocksPreferences {
 
 export type NavPreferences = Record<'files' | 'packages' | 'queries', boolean>
 
-interface PackagePreferencesInput {
-  message?: true
+export interface PackagePreferencesInput {
+  message?: boolean
   user_meta?: ReadonlyArray<string>
 }
 export interface PackagePreferences {
@@ -187,7 +188,10 @@ const bucketPreferencesValidator = makeSchemaValidator(bucketPreferencesSchema)
 export function validate(data: unknown): asserts data is BucketPreferencesInput {
   const obj = typeof data === 'string' ? YAML.parse(data) : data
   const errors = bucketPreferencesValidator(obj)
-  if (errors.length) throw new bucketErrors.BucketPreferencesInvalid({ errors })
+  if (errors.length) {
+    if (errors[0] instanceof Error) throw errors
+    throw new JsonInvalidAgainstSchema({ errors: errors as ErrorObject[] })
+  }
 }
 
 function parseActions(actions?: Partial<ActionPreferences> | false): ActionPreferences {